issue #189 merge oracle compilers
move extension classes

initial merge of oracle compilers

update unit tests part one

rename test cs file

remove commented / obsolete tests
