+++ QueryBuilder/Compilers/MySqlCompiler.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace SqlKata.Compilers
 {
     public class MySqlCompiler : Compiler
@@ -46,15 +44,6 @@ namespace SqlKata.Compilers
 
             return "LIMIT ? OFFSET ?";
 
-        }
-    }
-
-    public static class MySqlCompilerExtensions
-    {
-        public static string ENGINE_CODE = "mysql";
-        public static Query ForMySql(this Query src, Func<Query, Query> fn)
-        {
-            return src.For(MySqlCompilerExtensions.ENGINE_CODE, fn);
         }
     }
 }
