+++ QueryBuilder/Compilers/PostgresCompiler.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace SqlKata.Compilers
 {
     public class PostgresCompiler : Compiler
@@ -38,15 +36,6 @@ namespace SqlKata.Compilers
             }
 
             return sql;
-        }
-    }
-    public static class PostgresCompilerExtensions
-    {
-        public static string ENGINE_CODE = "postgres";
-
-        public static Query ForPostgres(this Query src, Func<Query, Query> fn)
-        {
-            return src.For(PostgresCompilerExtensions.ENGINE_CODE, fn);
         }
     }
 }
