+++ QueryBuilder/Compilers/SqlServerCompiler.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace SqlKata.Compilers
 {
     public class SqlServerCompiler : Compiler
@@ -170,15 +168,6 @@ namespace SqlKata.Compilers
             }
 
             return sql;
-        }
-    }
-
-    public static class SqlServerCompilerExtensions
-    {
-        public static string ENGINE_CODE = "sqlsrv";
-        public static Query ForSqlServer(this Query src, Func<Query, Query> fn)
-        {
-            return src.For(SqlServerCompilerExtensions.ENGINE_CODE, fn);
         }
     }
 }
